pyproject.toml
[project]
name = "mcp-server-feishu"
version = "0.1.0"
description = "MCP Server for Feishu/Lark"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "markdownify>=0.11.6"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
